long

+++ include/gfastats-gfa.h
@@ -854,7 +854,7 @@ public:
         unsigned int seqLen = sequence->length()-1;
         for (char &base : *sequence) {
 
-            unsigned int count = 1;
+            unsigned long long int count = 1;
             if(hc_flag && hc_index < bedCoords.size() && pos == bedCoords[hc_index].first) {
                 count = bedCoords[hc_index].second - bedCoords[hc_index].first;
                 ++hc_index;
